+++ tests/cases.js
@@ -13,7 +13,7 @@ module.exports = [
 	'--test-entry base/b-virtual-scroll/test',
 	'--test-entry base/b-tree/test',
 	'--test-entry base/b-sidebar/test',
-	'--test-entry base/b-slider/test',
+	'--test-entry base/b-slider/test --browsers chrome --device iPhone_11',
 	'--test-entry base/b-list/test',
 	'--test-entry base/b-window/test',
 	'--test-entry form/b-button/test',