+++ src/base/b-list/CHANGELOG.md
@@ -9,6 +9,16 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.123 (2021-01-15)
+
+#### :rocket: New Feature
+
+* Now the component implements the `iItems` trait
+
+#### :memo: Documentation
+
+* Improved documentation
+
 ## v3.0.0-rc.90 (2020-10-22)
 
 #### :boom: Breaking Change