+++ src/form/b-button/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.123 (2021-01-15)
+
+#### :rocket: New Feature
+
+* Implemented new API from `iAccess`
+
 ## v3.0.0-rc.88 (2020-10-13)
 
 #### :bug: Bug Fix